% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_extract_haplotype.r
\name{extract_haplotype}
\alias{extract_haplotype}
\title{Extract haplotype using Haploview}
\usage{
extract_haplotype(combined_gwas_result = NULL,
  output_path = file.path("~"), Haploview_file_path = NULL,
  Haploview_file_name = NULL, Haploview_file_extension = NULL,
  Haploview_file_named_sequentially_from = NULL,
  Haploview_file_named_sequentially_to = NULL)
}
\arguments{
\item{output_path}{An output path.}
}
\value{
combined_gwas_result or NULL if something missing.
}
\description{
The goal of search_genes is to extract haplotype blocks and append the results to combined_gwas_result.
}
\keyword{Haplotype_blocks}
