% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_farming_with_GAPIT.r
\name{farming_with_GAPIT}
\alias{farming_with_GAPIT}
\title{Farming with GAPIT}
\usage{
farming_with_GAPIT(dat = NULL, by_column = 1, start_column = 2,
  output_path, p_value_threshold = NA, p_value_fdr_threshold = NA,
  ld_number = 0, KI = NULL, CV = NULL, G = NULL, GD = NULL,
  GM = NULL, file.Ext.G = NULL, file.Ext.GD = NULL,
  file.Ext.GM = NULL, file.G = NULL, file.GD = NULL,
  file.GM = NULL, file.path = NULL, file.from = 0, file.to = 0,
  model = NULL, SNP.MAF = 0, PCA.total = 0,
  Model.selection = FALSE, SNP.test = FALSE, file.output = FALSE)
}
\arguments{
\item{dat}{An input dataset.}

\item{by_column}{The accession column.}

\item{start_column}{The start column index for traits.}

\item{output_path}{An output path.}
}
\value{
combined_gwas_result or NULL if something missing.
}
\description{
The goal of farming_with_GAPIT is to run GWAS with GAPIT.
}
\keyword{GAPIT,}
\keyword{GWAS}
