% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_generate_BLUE.r
\name{generate_BLUE}
\alias{generate_BLUE}
\title{generate BLUE function}
\usage{
generate_BLUE(dat = NULL, by_column = c(1, 2), start_column = 3)
}
\arguments{
\item{dat}{An input dataset.}

\item{by_column}{The accession column.}

\item{start_column}{The start column index for traits.}
}
\value{
blue or NULL if something missing.
}
\description{
The goal of generate_BLUE is to run Best Linear Unbiased Estimations.
}
\keyword{BLUE,}
\keyword{Pre-GWAS}
