% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_generate_BLUP.r
\name{generate_BLUP}
\alias{generate_BLUP}
\title{generate BLUP function}
\usage{
generate_BLUP(dat = NULL, by_column = c(1, 2), start_column = 3)
}
\arguments{
\item{dat}{An input dataset.}

\item{by_column}{The accession column.}

\item{start_column}{The start column index for traits.}
}
\value{
blue or NULL if something missing.
}
\description{
The goal of generate_BLUP is to run Best Linear Unbiased Predictions.
}
\keyword{BLUP,}
\keyword{Pre-GWAS}
