% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HAPPI.GWAS-package.R
\docType{package}
\name{HAPPI.GWAS-package}
\alias{HAPPI.GWAS}
\alias{HAPPI.GWAS-package}
\title{HAPPI.GWAS: HAPPI.GWAS}
\description{
HAPPI.GWAS is a package for GWAS.
}
\author{
\strong{Maintainer}: Yen On Chan \email{chanye@mail.missouri.edu} (https://orcid.org/0000-0002-3345-9614)

}
\keyword{internal}
