% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_search_genes.r
\name{search_genes}
\alias{search_genes}
\title{Search genes from gff files}
\usage{
search_genes(combined_gwas_result = NULL, output_path = file.path("~"),
  GFF_file_path = NULL, GFF_file_name = NULL,
  GFF_file_extension = NULL, GFF_file_named_sequentially_from = NULL,
  GFF_file_named_sequentially_to = NULL)
}
\arguments{
\item{output_path}{An output path.}
}
\value{
combined_gwas_result or NULL if something missing.
}
\description{
The goal of search_genes is to append genes to combined_gwas_result.
}
\keyword{Genes}
